[RW-7275][risk=no] rm old commands and docker usage from UI
