--- ui/docker-compose.yaml
@@ -1,28 +1,0 @@
-version: "3"
-services:
-  ui:
-    build:
-      context: ./src/dev/server
-    image: allofustest/workbench-dev-ui:local
-    user: ${UID}
-    working_dir: /w/ui
-    volumes:
-      - ..:/w:cached
-      - ~/.config:/.config:cached
-      - ~/.gsutil:/.gsutil:cached
-    command: yarn start --host=0.0.0.0 ${REACT_APP_ENVIRONMENT}
-    environment:
-      - GOOGLE_APPLICATION_CREDENTIALS=/w/ui/sa-key.json
-    ports:
-      - 127.0.0.1:4200:4200
-  tests:
-    build:
-      context: ./src/dev/server
-    image: allofustest/workbench-dev-ui:local
-    user: ${UID}
-    working_dir: /w/ui
-    volumes:
-      - ..:/w:cached
-    command: yarn test
-    ports:
-      - 127.0.0.1:9876:9876
